# Conflicts:
#	src/database/repositories/round.repository.ts

+++ src/database/repositories/deliverable.repository.ts
@@ -93,13 +93,32 @@ export class DeliverableRepository
         }
     }
 
-    public async delete(id: string): Promise<boolean> {
+    public async delete(query: FindQuery): Promise<boolean> {
         try {
+            if (!query || !query?.find) {
+                throw new Error('Please specify a query');
+            }
+
             const response: MongooseDeleteResponse = await this.model.deleteOne(
-                { id: id },
+                query.find,
             );
 
             return response.deletedCount === 1;
+        } catch (error: any) {
+            throw error;
+        }
+    }
+
+    public async deleteMany(query: FindQuery): Promise<boolean> {
+        try {
+            if (!query || !query?.find) {
+                throw new Error('Please specify a query');
+            }
+
+            const response: MongooseDeleteResponse =
+                await this.model.deleteMany(query.find);
+
+            return response.deletedCount > 0;
         } catch (error: any) {
             throw error;
         }
